Remove the session cache of the "RT at a Glance" portlet lists
The initial commit of "RT at a Glance" configurability (8b8e90d) added a
caching layer in the session, rather than hit the Attribute for the
preferences on each request.  Remove this caching, in order to ensure
that users who have not overridden the preference always get the most
recent global configuration.  This comes at the cost of an extra 1-2
Attribute queries on the front page, but as these are well-indexed, this
negligable performance impact seems worth it to obey the principle of
least surprise when the admin updates the global configuration.