lower number of POSIX::tzset calls
We call tzset twice per date to set new and restore old values. These
two calls on my laptop take 0.0014s. For ticket's history with 100
transactions (any txn, not only replies) it takes 0.14s to call tzset.

Instead we can compare TZ env to required value in RT::Date. If
values are equal we can skip tzset.

To further increase chances of equality we set TZ env to current
user's environment and clean it up when request ends. This part
is QUESTIONABLE as it probably messes with calls to system.